+++ src/alicechess/pieces/__init__.py
@@ -9,6 +9,7 @@ import typing as _t
 from alicechess.pieces.king import King
 from alicechess.pieces.pawn import Pawn
 from alicechess.pieces.piece import Piece
+from alicechess.position import BoardPosition
 from alicechess.utils import Color, PieceType, PromoteType
 
 # =============================================================================
@@ -34,16 +35,19 @@ def _make_simple_piece(piece_type, rows, columns=None) -> _t.Type[Piece]:
     class Subclass(Piece):
         _type = piece_type
 
-        def _check_start_position(self) -> bool:
-            if self._pos.bn != 0:
+        def is_at_start_pos(
+            self, pos: _t.Optional[BoardPosition] = None
+        ) -> bool:
+            bn, r, c = pos or self._pos
+            if bn != 0:
                 return False
             if self._color is Color.WHITE:
                 row = rows[0]
             else:
                 row = rows[1]
-            if self._pos.r != row:
+            if r != row:
                 return False
-            if columns is not None and self._pos.c not in columns:
+            if columns is not None and c not in columns:
                 return False
             return True
 
@@ -72,6 +76,4 @@ def make_promoted(
         PromoteType.KNIGHT: Knight,
         PromoteType.BISHOP: Bishop,
     }
-    return PROMOTE_CLASSES[promote_type](
-        pawn.id, pawn.color, pawn.pos, has_moved=pawn.has_moved
-    )
+    return PROMOTE_CLASSES[promote_type](pawn.id, pawn.color, pawn.pos)